feat: real lib kickoff
