+++ .gitignore
@@ -1,1 +1,2 @@
-NON_COMPLETE_RUST_WRAPPER/hello_world/target
+target
+*.lib
