+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "my-webui-application"
+name = "webui-rs"
 version = "0.1.0"
 build = "build.rs"
 edition = "2021"
