--- NON_COMPLETE_RUST_WRAPPER/hello_world/README.md
@@ -1,9 +1,0 @@
-# WebUI Examples - Rust
-
-```sh
-git clone https://github.com/alifcommunity/webui.git
-cd webui\examples\Rust\hello_world
-cargo clean
-cargo build
-cargo run
-```
