+++ build.rs
@@ -1,7 +1,4 @@
-
-
 fn main() {
-
     // Path to prebuilt WebUI static library
     println!("cargo:rustc-link-search=../../../build/Windows/MSVC/");
 
@@ -9,6 +6,6 @@ fn main() {
     println!("cargo:rustc-link-search=../../../include/");   
     
     // WebUI static library name
-    println!("cargo:rustc-link-lib=webui-2-static-x64");
+    println!("cargo:rustc-link-lib=webui-2-static");
     println!("cargo:rustc-link-lib=user32");
 }
